work

+++ backend/src/main/java/it/chalmers/gamma/group/service/GroupService.java
@@ -3,6 +3,7 @@ package it.chalmers.gamma.group.service;
 import it.chalmers.gamma.domain.IDsNotMatchingException;
 import it.chalmers.gamma.group.data.Group;
 import it.chalmers.gamma.group.data.GroupRepository;
+import it.chalmers.gamma.group.dto.GroupDTO;
 import it.chalmers.gamma.group.dto.GroupShallowDTO;
 import it.chalmers.gamma.group.exception.GroupNotFoundException;
 import it.chalmers.gamma.group.controller.response.GroupDoesNotExistResponse;
@@ -31,23 +32,25 @@ public class GroupService {
         this.repo.save(new Group(groupFinder.fromShallow(group)));
     }
 
-    public void editGroup(GroupShallowDTO oldGroup) throws GroupNotFoundException, IDsNotMatchingException {
-        Group group = this.groupFinder.getGroupEntity(oldGroup.getId());
-        group.apply(this.groupFinder.fromShallow(oldGroup));
+    public void editGroup(GroupDTO newEdit) throws GroupNotFoundException, IDsNotMatchingException {
+        Group group = this.groupFinder.getGroupEntity(newEdit);
+        group.apply(newEdit);
         this.repo.save(group);
     }
 
-    public void removeGroup(UUID groupId) {
-        this.repo.deleteById(groupId);
+
+    public void editGroup(GroupShallowDTO newEdit) throws GroupNotFoundException, IDsNotMatchingException {
+        Group group = this.groupFinder.getGroupEntity(newEdit);
+        group.apply(groupFinder.fromShallow(newEdit));
+        this.repo.save(group);
     }
 
-    public void editGroupAvatar(UUID groupId, String url) throws GroupNotFoundException {
-        Group group = groupFinder.getGroupEntity(groupId);
-        if (group == null) {
-            throw new GroupDoesNotExistResponse();
+    public void removeGroup(UUID groupId) throws GroupNotFoundException {
+        if(!this.groupFinder.groupExists(groupId)) {
+            throw new GroupNotFoundException();
         }
-        group.setAvatarURL(url);
-        this.repo.save(group);
+
+        this.repo.deleteById(groupId);
     }
 
 }
